Reorg Repository and Start Documentation
