+++ go.mod
@@ -3,7 +3,11 @@ module github.com/Fishwaldo/go-logadapter
 go 1.16
 
 require (
+	github.com/kr/pretty v0.2.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	github.com/sirupsen/logrus v1.8.1
 	github.com/stretchr/testify v1.7.0
 	go.uber.org/zap v1.19.1
+	golang.org/x/sys v0.0.0-20210809222454-d867a43fc93e // indirect
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
