+++ loggers/std/std_test.go
@@ -1,24 +1,23 @@
-package logadapter
+package stdlogger_test
 
 import (
 	"bytes"
 	"os"
 	"regexp"
 	"testing"
+
+	"github.com/Fishwaldo/go-logadapter/loggers/std"
 )
 
 func TestDefaultLogger(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
-	if logger.level != LOG_TRACE {
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
+	if logger.GetLevel() != stdlogger.LOG_TRACE {
 		t.Errorf("Can't Set Logging Level")
-	}
-	if logger.GetLevel() != LOG_TRACE {
-		t.Error("GetLevel Didn't return Correct Logging Level")
 	}
 }
 
-func captureOutput(l *StdLogger, f func()) string {
+func captureOutput(l *stdlogger.StdLogger, f func()) string {
     var buf bytes.Buffer
     l.Log.SetOutput(&buf)
     f()
@@ -26,8 +25,8 @@ func captureOutput(l *StdLogger, f func()) string {
     return buf.String()
 }
 func TestLogTrace(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
 	output := captureOutput(logger, func() { 
 		logger.Trace("Hello %s", "world")
 	})
@@ -37,8 +36,8 @@ func TestLogTrace(t *testing.T) {
 	}
 }
 func TestLogDebug(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
 	output := captureOutput(logger, func() { 
 		logger.Debug("Hello %s", "world")
 	})
@@ -49,8 +48,8 @@ func TestLogDebug(t *testing.T) {
 }
 
 func TestLogInfo(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
 	output := captureOutput(logger, func() { 
 		logger.Info("Hello %s", "world")
 	})
@@ -61,8 +60,8 @@ func TestLogInfo(t *testing.T) {
 }
 
 func TestLogWarn(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
 	output := captureOutput(logger, func() { 
 		logger.Warn("Hello %s", "world")
 	})
@@ -73,8 +72,8 @@ func TestLogWarn(t *testing.T) {
 }
 
 func TestLogError(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
 	output := captureOutput(logger, func() { 
 		logger.Error("Hello %s", "world")
 	})
@@ -97,8 +96,8 @@ func TestLogFatal(t *testing.T) {
 }
 
 func TestLogPanic(t *testing.T) {
-	logger := DefaultLogger()
-	logger.SetLevel(LOG_TRACE)
+	logger := stdlogger.DefaultLogger()
+	logger.SetLevel(stdlogger.LOG_TRACE)
 	defer func() {
 		if err := recover(); err == nil {
 			t.Errorf("Log Panic Recovery Failed")
